oops

+++ index.js
@@ -14,13 +14,15 @@ function CreateShortenURL(url) {
         shortURL: sha(link).slice(0, 7)
     }
     var sURL = SavedURL.shortURL;
-    !existsSync('./ShortData/') || mkdirSync('./ShortData/');
+    if(!existsSync('./ShortData/')) return mkdirSync('./ShortData/');
     if (existsSync(`./ShortData/${sURL}.json`)) {
-        !JSON.parse(readFileSync(`./ShortData/${sURL}.json`, 'utf8')).sURL == sURL) || sURL = sha(link).slice(0, 8);
+        if (!JSON.parse(readFileSync(`./ShortData/${sURL}.json`, 'utf8')).sURL == sURL) {
+            sURL = sha(link).slice(0, 8);
+        }
     }
     writeFileSync(`./ShortData/${sURL}.json`, JSON.stringify(SavedURL), (err) => { console.log(err) });
     //return "https://ddc.gg/t/" + sURL;
-    return "http://localhost:8069" + sURL;
+    return "http://localhost:8069/" + sURL;
 };
 app.get('', (_req, res) => {
     res.sendFile(__dirname + "/index.html");
